minor tweaks
git-svn-id: https://svn.code.sf.net/p/jackcess/code/jackcess/trunk@399 f203690c-595d-4dc9-a70b-905162fa7fd2

+++ src/java/com/healthmarketscience/jackcess/query/Query.java
@@ -319,7 +319,7 @@ public abstract class Query
         comboJoins = new ArrayList<Row>();
         comboJoinMap.put(key, comboJoins);
       } else {
-        if(comboJoins.get(0).flag != join.flag) {
+        if((short)comboJoins.get(0).flag != (short)join.flag) {
           throw new IllegalStateException(
               "Mismatched join flags for combo joins");
         }
